Review bug fix

+++ app/src/main/java/com/bupp/wood_spoon_eaters/bottom_sheets/reviews/ReviewsAdapter.kt
@@ -43,9 +43,9 @@ class ReviewsAdapter : ListAdapter<CommentAdapterItem, RecyclerView.ViewHolder>(
                 reviewsItemName.text = data.eater?.firstName
                 reviewsItemComment.text = data.reviewText ?: ""
                 data.eater?.thumbnail?.let { reviewsItemIcon.setImage(it) }
-                val rating = data.rating ?: 0
-                reviewsItemRating.isVisible = rating > 0
-                reviewsItemRating.setRating(rating)
+                val rating = data.rating ?: 0.0
+                reviewsItemRating.isVisible = rating > 0.0
+                reviewsItemRating.setRating(rating.toInt())
             }
         }
     }
